Commit remarks in homework 4.2
