+++ src/components/Cards/Cards.jsx
@@ -6,9 +6,9 @@ import { EndGameModal } from "../../components/EndGameModal/EndGameModal";
 import { Button } from "../../components/Button/Button";
 import { Card } from "../../components/Card/Card";
 import { useEasyMode } from "../../contexts/easyModeContext/UseEasyMode";
-import { useLeaders } from "../../contexts/leaderContext/UseLeaders";
 import { useUser } from "../../contexts/userContext/UseUser";
-import { getLeaders, postLeader } from "../../api";
+import eyeIcon from "../../icons/eye.svg";
+import force2Icon from "../../icons/force2.svg";
 
 // Игра закончилась
 const STATUS_LOST = "STATUS_LOST";
@@ -46,10 +46,10 @@ function getTimerValue(startDate, endDate) {
  */
 export function Cards({ pairsCount = 3, previewSeconds = 5 }) {
   const [isLeader, setIsLeader] = useState(false);
-  const [setError] = useState(null);
+  // const [setError] = useState(null);
   const { isEasyMode, selectedLevel } = useEasyMode();
-  const { leaders, setLeaders } = useLeaders();
-  const { user, setUser } = useUser();
+  // const { setLeaders } = useLeaders();
+  const { setUser } = useUser();
   // Если игорок выбирает легкий уровень с 3 попытками, в attempts организован счетчик этих попыток
   const [attempts, setAttempts] = useState(isEasyMode ? 3 : 1);
   // В cards лежит игровое поле - массив карт и их состояние открыта\закрыта
@@ -88,46 +88,7 @@ export function Cards({ pairsCount = 3, previewSeconds = 5 }) {
     if (isEasyMode) {
       setAttempts(3);
     }
-    if (isLeader) {
-      postUserLeaderboard();
-    }
   }
-
-  const postUserLeaderboard = async () => {
-    if (user) {
-      const timeUser = timer.minutes * 60 + timer.seconds;
-      try {
-        return postLeader({ user, timeUser });
-      } catch (error) {
-        console.error(error.message);
-        if (error.message === "Failed to fetch") {
-          setError("Ошибка соединения");
-          return;
-        }
-      }
-    }
-  };
-
-  // Можем ограничить запись результата пользователя вхождением этого показателя
-  // в какой-то интервал, допустим, в тройку или десятку лучших
-  const checkLeaderboard = () => {
-    let result = false;
-    const bestResults = 10;
-    const checkCount = leaders.length < bestResults ? leaders.length : bestResults;
-    const timeUser = timer.minutes * 60 + timer.seconds;
-    // Отсортируем результаты лидерборда в порядке возрастания времени
-    const sortLeaders = leaders.sort((a, b) => a.time - b.time);
-
-    for (let i = 0; checkCount - 1; i++) {
-      if (timeUser <= sortLeaders[i].time) {
-        result = true;
-        break;
-      }
-    }
-    // Пока не будем ограничивать запись результатов в лидербор
-    result = true;
-    return result;
-  };
 
   /**
    * Обработка основного действия в игре - открытие карты.
@@ -159,10 +120,9 @@ export function Cards({ pairsCount = 3, previewSeconds = 5 }) {
 
     // Победа - все карты на поле открыты
     if (isPlayerWon) {
-      // Если выполняются условия: не лёгкий режим, 3-й уровень сложности
-      // и (опционно) вхождение результата в Х-топ, предоставляется возможность записи
-      // результата в лидерборд
-      if (checkLeaderboard() & !isEasyMode & (selectedLevel === 9)) {
+      // Если выполняются условие: 3-й уровень сложности
+      // предоставляется возможность записи результата в лидерборд
+      if (selectedLevel === 9) {
         setIsLeader(true);
       }
       finishGame(STATUS_WON);
@@ -212,7 +172,6 @@ export function Cards({ pairsCount = 3, previewSeconds = 5 }) {
         if (attempts === 0) {
           finishGame(STATUS_LOST);
           setAttempts(3);
-          // setUser("Пользователь");
           return;
         }
       }
@@ -264,19 +223,6 @@ export function Cards({ pairsCount = 3, previewSeconds = 5 }) {
     }
   }, [attempts]);
 
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const response = await getLeaders();
-        setLeaders(response.leaders);
-      } catch (error) {
-        console.error(error);
-        setError("Ошибка при получении списка лидеров");
-      }
-    };
-    fetchData();
-  }, [setLeaders, setError]);
-
   return (
     <div className={styles.container}>
       <div className={styles.header}>
@@ -299,6 +245,12 @@ export function Cards({ pairsCount = 3, previewSeconds = 5 }) {
               </div>
             </>
           )}
+        </div>
+        <div>
+          <img src={eyeIcon} className={styles.achievement} alt="eye" />
+        </div>
+        <div>
+          <img src={force2Icon} className={styles.achievement} alt="force" />
         </div>
         {status === STATUS_IN_PROGRESS ? <Button onClick={resetGame}>Начать заново</Button> : null}
       </div>
