+++ src/components/Cards/Cards.module.css
@@ -94,3 +94,7 @@
   flex-wrap: nowrap;
   justify-content: flex-end;
 }
+
+.achievement {
+  display: none;
+}
