+++ src/components/EndGameModal/EndGameModal.module.css
@@ -75,6 +75,37 @@
   text-align: center;
   background-color: #fff;
   border: none;
-  /* color: #999999; */
   color: #000;
+}
+
+.modalUser {
+  display: flex;
+  flex-direction: row;
+  gap: 20px;
+}
+
+.nameSave {
+  width: 150px;
+  height: 45px;
+  border-radius: 12px;
+  background: #7ac100;
+
+  color: #fff;
+  text-align: center;
+  font-variant-numeric: lining-nums proportional-nums;
+
+  /* Pres → Caption S */
+  font-family: StratosSkyeng;
+  font-size: 24px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 32px; /* 133.333% */
+
+  border: none;
+
+  cursor: pointer;
+}
+
+.button:hover {
+  background: #7ac100cc;
 }
