+++ src/pages/LeaderboardPage/LeaderboardPage.module.css
@@ -59,7 +59,7 @@
 .userName,
 .userTime,
 .emptySpace {
-  font-family: StratosSkyengs;
+  font-family: Montserrat1;
   font-size: 24px;
   font-weight: 400;
   line-height: 32px;
